ss2/requirements.txt
pymysql>=1.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0